Remove explicit dependency on qemu-kvm. Add explicit dependency on libvirt-daemon-kvm (on x86).
With the libvirt backend, we don't invoke qemu/qemu-kvm directly (with
the exception of virt-rescue).

If we now assume Fedora 18+, then there's no reason to depend on a
version of qemu, since all qemu versions contain virtio-scsi.

However on x86, add a dependency on libvirt-daemon-kvm so that
qemu-kvm is pulled in indirectly.

Thanks Dan Berrange, Peter Robinson.